requirements.txt
haystack-ai>=2.0.0
requests>=2.25.0
python-dotenv>=1.0.0